If no project to dsbt-setup, load all artifacts.
This is a somewhat more useful behavior: if no project name is
given to the command 'dsbt-setup', then all the artifacts of all
the projects known in the build will be reloaded. This can be useful
in case the full set of artifacts is used to test a project that was
/not/ in the dbuild build file: everything known will be reloaded
(dsbt-setup should still work as intended).