Major config reorganization
The config code was previously relying on a manual serialization and
deserialization; that suffered from limited error checking, and much
code replication.

In this commit, all that code has been drastically simplified, and
replaced by much simpler calls to a customized version of the "Jacks"
library, which in turn relies on Jackson.

At this time all the refactoring should be complete: all the code
compiles, and the metadata test suite completes succesfully. Small
refinements and more testing will follow.

The private (and patched) copy of the Jacks library should be
replaced with the mainline, when (if) the relevant patches are
included upstream.