Ask the build system to expand its build defaults
Prior to the metadata refactoring, the 'extra' field was only applied to
sbt; the defaults were merged at the time of the config loading. The
default sbt version was inserted at this point, grabbing it from the
sbt defaults.

After the reorganization, however, the 'extra' component applies to
multiple build systems. The SimpleBuildActor now demands expanding
the defaults to each build system, by sending messages to the
corresponding builder. That also allows for a better encapsulation
of the build-specific aspects into each BuildSystem.