Fix git checkout of repositories
In case the same project is reused, an existing Git repository
is reused. However, that was done with git fetch + git checkout,
where checkout uses a local branch. Unfortunately git fetch does
/not/ update local branches. To fix that, we explicitly check out
the remote branch instead (which is the one we are interested in
anyway). If the supplied fragment is not a remote branch, we
check whether the supplied fragment is an existing tag or an
existing commit.