Removed plugin dependency on akka actors
This commit reorganizes packages so that the sbt plugin will no
longer cause the akka actors jar to be added to the classpath.

The project "logging" is now split into "logging" (base project)
and "actorLogging", which includes actor-base functionality.
Similarly, some non-actor related functionality previously in
project "projects" is now in project "core", for use by the
plugin project.

This refactoring solves a problem that prevented recent versions
of Akka from compiling with dbuild. In detail, the Akka project
definition files live in an "akka" package, and may use sbt's
"config" facility. Having akka-actor on the classpath at this
point causes Akka's "akka.config" package to be on the
classpath, causing troubles. Removing the dependency solves
this issue.

A further colliding "package config" was previously removed
from dbuild as part of the large config refactoring in commit
1274d06212dcba89c9664e893df06d0057331439.