Use sbt options properly while extracting
A puzzling failure of the extraction phase in getting ALL of the
dependencies led to a painful chase, at the end of which I
discovered that the sbt options were passed to the sbt runner
during build, but not during extraction; that affected the
list of detected dependencies. Now fixed.