more work towards reorganizing the settings
It's really tricky. Appending settings, rather than updating them
in place, causes a premature update, which causes Ivy errors or in
any case missing artifacts (as they will be rematerialized later).

Yet, I need to be able to append, as in certain projects (sbt,
for example), there is no built-in publishTo definition that we
can overwrite....!!