Fix & refactoring required for the 'directory' extra option
We had a situation where the Sbt build system would use the
'directory' option if present; then LocalBuildRunner would
not use it (rematerializing therefore artifacts in the
checkout root, rather than the project root); and then the
build routine would use and need 'directory' again.

The various occurrences are now unified under the BuildRunner
method projectDbuildDir(), which asks each build system to
supply the proper location of the project-specific ".dbuild"
directory, in accordance to whatever is in the build system's
extra options.