Restrict extraction only to requested subprojects
On occasion, cycles may appear among selected subprojects of
some of the dbuild projects. By using the 'projects' extra
sbt option, in theory it should be possible to exclude the
offending subprojects. However, until now, all of the
subprojects' dependencies were extracted and checked: that
made it impossible to avoid such cycles by excluding
subprojects.

With this commit, dependency extraction is filtered
through the list of requested projects (if present).
Some of those cycles can therefore now be avoided.