Fixed incorrect testing of subprojects
The wrong scope was applied to the test task during
testing of sbt projects. As a result, if an sbt project
had 20 subprojects, the result would be that the same
root subproject would be tested 20 times.

That is not only incorrect, but may actively cause
problems when certain subprojects have been explicitly
excluded by using the 'projects' extra option.

Now fixed.