Fixing scalaBinaryVersion settings
The scalaBinaryVersion settings must be identical to the
scalaVersion strings; otherwise, sbt might fetch from
Maven Central or other external repositories projects
that have not been compiled with our version of Scala,
but happen to be binary compatible according to the
scalaBinaryVersion string. That would make the testing
of the chain of projects we are interested in to be
not repeateable, and incomplete.