Extending fixModule with support for explicit artifacts
The previous implementation of fixModule assumed in
general that the module name and the artifact name
will coincide. In certain cases, however, that is
definitely not the case. Specifically, the
compiler-interface sbt project is published as a
module "compiler-interface" containing two artifacts named
"compiler-interface-bin" and "compiler-interface-src"
(no classifiers).

The mismatch will cause troubles either in fixModule,
or during the Ivy resolution: they cannot work together.

The alternatives are:

1) rename the artifact compiler-interface-bin to
compiler-interface during the sbt publishing, and
suppress the additional artifact, appending an
explicit:

set artifact in (compileInterfaceSub, packageBin) := Artifact("compiler-interface")
set publishArtifact in (compileInterfaceSub, packageSrc) := false

2) slightly alter fixModule so that, in case explicit artifacts
are added to the ModuleID, the artifact names in the
rematerialized repository are /also/ matched agains those
explicit artifacts. The client code will look like:

libraryDependencies ++= Seq(
    "org.scala-sbt" % "compiler-interface" % "1.2.3" artifacts(Artifact("compiler-interface-bin")),
    ...

3) change fixModule quite more in depth. Rather than
trying to match modules and artifacts manually, we
should really be asking Ivy whether the module can
be resolved agains our local repo. That would be
rather more general, since then all the extra attributes
and so on would be taken into account directly by Ivy.
This option require using Ivy programmatically, which
is not too difficult but requires quite a bit of
additional code.

This commit implements 2).