Adding Git repositories caching
A single copy of each Git repository, matching a URI without
final fragment, is kept in target/clones. All the branches
in the cached repository are set to track the corresponding
remote ones, and the cached repository is updated with a
pull before each use.

All further local copies are made using a clone -l, meaning
that the repositories no longer need to be unnecessarily
duplicated (hardlinks are used), and that the actual
cloning is now lightning fast.