Adding initial support for subproject publishing
Among the strings describing the projects that should be
deployed to a specific repository, it is now possible to
insert records that list only certain subprojects that
should be published within a certain project.

The additional config serialization/deserialization logic
is in Config.scala. This is for the configuration file
only, there is no code that parses these new records.