Propagate upward the map project -> artifacts
The information concerning which sbt subproject builds
which artifacts is now propagated further up the call
chain, up to buildStuff() in DistributedRunner.

If we want to make use of this information, we can
just now change printResult() and makeBuildResult(),
and tweak the structure of BuildArtifacts accordingly.

We will then have the information necessary to publish
only a subset of subprojects during deploy.