Making good use of the new projects list
Using the new inter-project dependency analysis, a new list
of projects is calculated during extraction; it is closer to
what is required in order to be able to split projects into
multiple parts, and to deal properly with multiple subprojects
and their dependency.

This list, obtained during extraction, is now made to bubble
up to the build stage, where it replaces the previous list of
subprojects, which was blindly taken out of the "extra" sbt
configuration section.

Next step will be to rearrange the inter-project resolver,
and probably also to disable the "aggregate" setting on all
subprojects.