Added alternate version of interproject patching
There are two basic possible approaches:
1) disabling the inter-project resolver altogether,
publishing subprojects in the local publish repo
one at a time, and adding the local publish repo
to the list of resolvers, or
2) trying (in a rather hackish way) to patch the
interproject resolver so that the subprojects that
are explicitly excluded are removed.

This commit adds the second version; although less
clean, it may be the only way to make it work with
setupcmd, where aggregate and standard interproject
dependencies cannot be disabled (otherwise the end
user would have to recompile each of the possibly
many projects manually).

This thing is a /huge/ can of worms. But we are
(painfully) getting there. There seems to be
no other shortcut that can make things work
properly without some pain.