Conforming dbuild-setup to the new order of things
The command dbuild-setup (should) now be able to deal with
explicit subproject exclusions, and project splitting.
The way in which this is done is by fetching the
RepeatableDistributedBuild from the specified uuid passed
to dbuild-setup; from there, the RepeatableProjectBuild
is obtained by matching against the project name argument
to dbuild-setup (now no longer optional), and the
calculated list of subprojects (including exclusions etc)
is also available in there.