Local repo and target dir are now versioned, plus code reorg
The target directory and the directory inside ~/.dbuild that
contain the local repository are now versioned; useful since
the metadata format in the repository has changed.

Additionally, the "Default.scala" managed source lives now
in the "d-repo" project, together with ProjectDirs. That
allows for a consolidation with the paths in "Repository.scala",
avoiding inconsistencies.

Also in this commit, the removal of unnecessary duplication
in packaging.scala.