Custom scalaInstance, for compliance with xml modularization
This commit removes the hardcoded list of jars used to
recreate a scala lib directory, as well as the reference
to scalaHome, which is no longer fully operational in 0.12.4
with 2.11.x.

Instead, the local-repo is now scanned for jars, which are
then copied considering a blacklist (taken from sbt 0.13).
In place of scalaHome, a custom scalaInstance is created
instead.