Adding support for Global Build Options
This commit adds a new (still undocumented) section to the
dbuild configuration file. The global section "build-options"
can be used to define options that affect the build of all
the projects; its content is propagated to DistributedBuildConfig,
and to each RepeatableProjectBuild (and SbtBuildConfig).

A planned application is controlling for the entire dbuild run
the ability to enable/disable cross-versioning and the
kind of scalaBinaryVersion that should be used. This is
extremely useful in order to detect libraries that are
dependencies of some of the projects, and forcing the build
to abort if they are missing, rather than picking a random
compatible version from Maven.