Addressed an incompatibility between "full" and explicitArtifacts
This workaround allows "full" cross-versioning to become finally
the default, which allows us to detect required Scala dependencies
that may not be in a dbuild configuration file.

If you need to issue a release, you will want to add the line
  build-options:{cross-version:standard}
in order to restore the normal cross-version level of the
various projects, prior to publication.