The Scala compiler is always needed with sbt
When using the sbt build system, we have an implicit needed
dependency on the scala-compiler jar, since it is fetched and
turned into a Scale home. Until now, the scala-compiler was
always provided by the Scala build system whenever a project
included scala-library as a dependency, but that is no longer
the case when scala-library is provided via the Ivy build
system. We need to add the scala-compiler.jar explicitly,
therefore, to the list of dependencies.