Changes to the scalaHome set up stage
The Ivy build system publishes (or re-publishes) its artifacts
currently using an Ivy layout. Therefore, the routines that
previously regenerated the directory used to set up scalaHome
need to be changed, in order to cater for Scala libraries
published in Ivy format, Maven format, or possibly a mixture
of the two.