A lot more Ivy configuration-related madness
Ivy configurations are now handled appropriately, and the
dependency rules are set more appropriately. One bit that
is still missing is supporting the "optional" config, which
can be done by resolving once more separately that config,
extracting the artifacts that have been fetched (and are
therefore optional), and adding to them the optional config
while creating the new ivy.xml (note that the same artifact
can be in multiple configurations).

Also still missing (but closer) is supporting set-version;
right now the artifacts will be republished with the same
version number they were fetched with.