Even more Ivy hammering. Now finally things work properly!
This commit aligns the version of scala-* libraries to that of
scala-library; although that should not be necessary in general,
the compilation sbt uses the same version for all Scala libraries
and gets very confused if that is not the case.

Also added in this patch is a fix to configurations; the sbt boot
requires the "default" configuration, and won't work with the
usual "compile". Therefore, we add further mappings to make things
work in that case as well.

The code needs (much) cosmetic polishing, but it seems to be fully
working at this time.

Possibly future improvements:

1) add a separate resolution stage using only the "optional"
configuration. That would detect which (Maven) libraries are
in the optional scope; they can then be added as optional
artifacts when creating the new ivy.xml, so that the client
code that should ask for optional libraries will find them
even with the ivy.xml of the republished artifacts. But this
only affects those (rare) clients, so it's not a real priority.
Note that jline (hence jansi) are optional dependencies in
scala-compiler. By default, dbuild will leave them out (which
is a good thing, otherwise sbt-republish gets confused), but
in theory they should be included in the ivy.xml as optional
dependencies.

2) add a further list of configurations that apply to the
project config (same level as main-jar, etc). These configurations
can then be used during resolution/dependency detection, in case
some obscure library uses non-standard Ivy configurations.
They could also offer an alternative to the optional stuff
described above (which would be better, being automatic
during resolution, though)

3) clean up the IvyMachinery code, which has an ugly
two-stage piece of code that should be refactored and
split properly.