Fixes and changes to DeployBuild
Since it is now called also when some of the projects failed,
it must check against the BuildOutcomes of the various projects,
and also be able to continue when one of the deployments fails.

Also renamed DeployElement and related classes to
SelectorElement, since I will shortly use it for notifications
as well.