Autowrapping of lonely strings into arrays in JSON config files
In many parts of the dbuild configuration file, lists of strings
are requested, with the result that the user has to remember which
options take single strings, which lists, brackets are forgot, etc.

This commit adds an automatic wrapping feature, by which every
time dbuild expects a list of strings, a single string can also
be supplied in the configuration file: the lonely string will
automatically be wrapped into a list. During serialization, the
opposite will also take place.