Partial solution to issue #35 (support github pull requests)
See issue #35
Specifying a url like ".....git#pull/NNN/head" should work now.

The drawback is that, for projects that have a very large
number of pull requests like Scala, a correspondingly large
number of commands are issued via Process(), resulting in
several thousands invocations to the command line git utility,
taking an atrociously long time.

As an interim solution, I check the uri fragment, and only
add the github PR refspec if the uri begins with "pull/".
So, the feature is added (albeit with a penalty for
usage), while leaving the previous behavior unaffected.

However, one run with a "pull/" uri fragment from a given
repository will be enough to "poison" the local clone
with thousands of tracking branches.

This issue remains open for the moment; a better solution
will come by moving to JGit, which is reportedly
very fast, and a much better approach than invoking
git via Process(). There are also other optimizations
that can be applied to the current Git cloning scheme.