Long refactoring, to bring buildOptions all the way to expandConfig
Since some of the global build options may affect extraction,
and specifically a default sbt version is in there, it is now
necessary to carry a copy of the build options around together
with the ProjectBuildConfig, all the way around extraction-
related routines. We do that with a new class ExtractionConfig.

The sbt build system is now able to find the default version
from the config file, rather than having to rely on whatever
version happens to be currently in use for dbuild.