Checking for missing deps among ModuleRevisionIds
This commit fixes an incorrect detection of
missing dependencies. A previous fix in 0.6.3,
which was meant to address the detection of
the current project artifacts as being not
part of dependencies, exposed the fact that
inter-project dependencies were not properly
recognized, and could have been incorrectly
detected as missing.

This commit properly fixes the situation, by
explicitly looking into the list of artifacts
published by the current project (actually,
by the list of selected subprojects), via
their list of ModuleRevisionIds.