Better handling of cycle diagnostics
prepareLog() normally trims stack traces to the first few lines.
However, the cycle detection logic within the graph submodule
may need to return a longer diagnostics; with this commit we
handle the CycleException specially, printing its extended
diagnostic separately, within prepareLog.

TODO: in SimpleBuildActor a cycle is not caught; this means
that, if a cycle is detected, an Unexpected outcome is generated,
and no notifications are sent, which is ungood. We must catch
that case instead, and process afterTasks as normal even in that
case.