Fixing the generation of scalaHome
This patch is a bit more robust of the one temporarily
applied. In particular, the rationale is:

- All artifacts in organization "org.scala-lang.modules"
are included. This is necessary since further modules
will be spawned shortly, and by definition that
organization will only be used for modules extracted
from the Scala compiler source (should that no longer
hold in the future, a further black list can be
introduced). For instance, scaladoc should be shortly
extracted from the main source, but the presentation
compiler relies in turn on scaladoc, therefore that
component needs to be extracted as well at some point.

- The underlying idea when using mavenJarFile and
ivyJarFile is that the main jar corresponding to a
given artifact needs to be in a very precise
location; if it is not there, an error condition
occurred, and we should be made aware of it. Hence,
scanning a subtree would be a less robust solution.

- The reason why retrieveJarFile() was not finding
scala-xml and the parser combinators was that
the scala artifacts were originally not cross-
versioned; that is no longer true now that modules
are present. Adding back to them the cross
suffix solves the issue at hand.