Adding more control over time out conditions, better notifications
This commit adds a bunch of watchdogs during extraction/building;
separate timeout durations can be selected for the entire extraction
phase, for each extraction, for each building, for extraction plus
building, and for the dbuild run as a whole.

The general one (which should never happen) will result in an
UnexpectedOutcome; all the other timeouts will be properly captured
and passed to the notification system, with informative messages.

Fixes #52