Deserialize SeqString as a Seq, rather than as an array
The previous deserialization code for SeqString was
implicitly creating a WrappedArray[String], which ends
up having a sha1 UUID that is different from that of
the original Seq (actually a List). Although the code
appeared to work as expected, the differences in
UUID caused drepo to be unable to parse the repository;
dbuild-setup was similarly unable to reload its context.

Now fixed; this commit closes #54.