Split project metadata into two, closes #57
The metadata of each project used to be published only
after the project completed its build successfully.
However, that made it impossible to debug a build using
the 'dbuild-setup' sbt plugin, even if all the
dependencies were already available.

This commit reorganizes the publication of project
metadata, which is now split into two. The
RepeatableProjectBuild portion is published
as soon as all of the project dependencies are
ready, and we attempt to build the project for
the first time. This information is already
sufficient to allow dbuild-setup to recreate
the necessary debugging environment.

Once the build completes successfully, conversely,
the artifacts information is written to the
repository separately, as a Seq[BuildSubArtifactsOut].

CAVEAT: normalizedProjectName() will currently get
confused if the git clone dir name is the same as the
name of a subproject, in which case it will mistake it
for a default root subproject name, when dbuild-setup
is used. More logic will be necessary to recognize
this case, see #66.