Adding "extraction-compiler" option and docs.
The "extraction-compiler" option can be used to select which
compiler should be used to extract the project dependencies.
It can be useful in case the project's default compiler belongs
to one Scala family (for example 2.10), but its definition adds
libraries for a different compiler, which is the one that we need.

In this case, the dependency would not be "seen", since extraction
by default takes place using whatever compiler is already defined
in the project.

This option allows the user to override that default,
selecting a compiler for extraction that is in the same
compatibility family as the one that we are interested in.