Split d-projects, factoring out the parts that deal with actors
This is done so that we can have a dependency on Extractor and
LocalBuildRunner (using "d-projects") without having a dependency
on actor-specific code, which we skip when compiling with 2.10
(in order to generate a 0.13 version of the plugin).

This is necessary since we must not pull in by mistake akka-actor
together with the plugin: if we do, akka compilation will fail
because of a naming conflict between the akka-actor lib and
a method defined in the akka build file.