Support for cross-versioning renaming in modules
Modules are compiled using their own "build.options"
specification, and are compiled in isolation.
However, the artifacts that will become part of the
rearranged resulting set need to have names that
are in compliance with whatever "cross-version"
option was specified by the user for the entire
configuration file.

This commit adds the necessary rewriting code.