Adding support for Scala "modules" generated by the ant task
The rewriting becomes quite hairy. The ant-generated artifacts
are manually inspected, and the cross-version suffix is
obtained directly from the path of the pom files.

Once the cross-version suffixes are known, the ant-generated
artifacts are split into two groups: one for the "real" core
(without cross versions), and additional modules (which are
cross-versioned).

The latter are then processed together with the artifacts
generated by nested projects, in terms of cross-version
patching and version rewriting, and rewiring in the pom files.