Add a "test" build system; "nil" now does really nothing
The build system "nil" was previously used for testing;
it would just fail every now and then, and do nothing
otherwise.

This functionality (for testing) is now available in the
"test" build system. The "nil" build system and resolver
now do really nothing; the resolver in particular will
be used as a default by future build systems that
require no source files.