Ivy is not my favorite piece of software
The artifact descriptions embedded in dependencies have an
independent name from the name of the dependency itself;
of course, there is no facility to duplicate easily either
a module descriptor, or a dependency descriptor, or an
artifact descriptor. So it's a matter of digging into the
Ivy source code, finding out what bits of information
should be in each, and trying to replicate all of those
structures, recursively, in order to adjust them.