Fix to logging system; tangentially related to #76
This commit implements an orderly shutdown of the
actor-based logging system. Previously, the actors
would just be stopped or killed half-way through,
causing parts of the logs to be lost. The postStop()
of LoggerFileWriteActors was never called,
incidentally.

Now a proper traversal of the actor hierarchy is
implemented, ensuring that all logs are properly
written to the console, as well as written to
the relevant files.

This fix is a consequence of investigation triggered
by #76, but is not a fix to it (the matter is
still open, and seems related to sbt's behavior).