Better capture of stack traces
In certain cases, exceptions were printed via
a printStackTrace(), which prints directly to
stderr, bypassing the log files. That could
cause the trace to appear on the console out
of place, or to be lost altogether.

This is related (and may possibly solve) #76,
although a bit more testing is needed.