Remove external dependencies from the "assemble" deps
The "assemble" build system builds each of its parts
independently, without interaction with the remaining
projects in the same dbuild configuration file.

As a consequence, the resulting assembled project
should not be dependent on other artifacts generated
by other projects, at the same level as the aggregate
(or above).

We don't have the infrastructure to do an extensive
checking for this condition, and "assemble" is a
sort of transitory hack anyway, therefore for the
time being we just strip away the external
dependencies, issuing a warning during extraction.
This is more than adequate for those dependencies
that are only needed at compile time (for instance,
scalacheck is needed during the compilation of
some Scala modules, but we cannot use a scalacheck
built by dbuild, as we don't have a full Scala
compiler yet. However, there is no danger of
library conflicts, as scalacheck is not actually
used by those modules at runtime).