MUCH better handling of Git repositories. Also: jgit supported
This commit completely reorganizes the git support infrastructure,
making it considerably faster and adding optional jgit support.
The number of git invocations is now reduced, by using specially
crafted refspecs that allow "fetch" to replicate all of the
remote tags, branches, etc, in a single operation.

The references of the form "pull/nnn/head", for Github pull requests,
are now fully supported and involve no particular speed penalty
during regular use.

Support for jgit has also been added. Since jgit has no support for
hardlinking, however, additional copies of all clones will be
made, which is both slower and uses more disk space. Because of
this reason, jgit is not the default resolver for "git://" schemes.
It can be selected, however, just by specifying a URI of the form
"jgit://" in one or all the projects.

This commit closes #48, and closes #35.