Move git support to a separate package
The git resolver uses jgit, but jgit may also be used by sbt.
The support-default project is a dependency of the sbt plugin,
and jgit changed packages structure in a non-compatible way,
meaning that either sbt or the dbuild sbt plugin may be
left unable to find their correct jgit classes, after
Ivy completes its resolution and picks one of the two.

This commit moves jgit to a separate project, which does not
get included with the dbuild sbt plugin, thereby removing
the conflict.