Restore user-repeatable configuration, other improvements
The root of the configuration file has been for a while
now no longer a DistributedBuildConfig, but rather a
DBuildConfiguration, which also includes the GeneralOptions.

Although the latter have no influence on the repeatability
of the build, saving only the RepeatableDistributedBuild
is insufficient, as users won't be able to use it as-is
to repeat the build. Further, a RepeatableDistributedBuild
includes ProjectConfigAndExtracted objects, and with the
schema strongly enforced those won't do in any case as
a legal dbuild configuration file.

This commit restores user repeatability of builds, by
creating a new DistributedBuildConfig after extraction,
which can now again be used as-is to repeat the build.

In the process, added warnings to Ivy's SNAPSHOT
resolutions, so that if the snapshot targets change
a warning is issued.