More solid renaming, in case multiple instances of the same dir are around
Sometimes it may happen that we have a directory that has been
renamed to "delete-...", but a new directory with the same name
has been created. If this second directory were to be deleted
as well, a collision would occur.
This commit takes care of locating a suitable destination name
for the renaming, even in that unlikely case.