Add the ability to build a single target project
A target project can now be specified on the command
line; dbuild will build just that project and its
dependencies, skipping the rest.

Extraction (which is a quick step anyway) will still
be performed on all projects, since this information
is required by dbuild in order to calculate the
graph of dependencies.