Changes to transitiveDependencyUUIDs
There is actually no need to recalculate the transitive
closure of the dependencies with transitiveDependencyUUIDs:
the list of dependencies is already transitive.

Now simplified to dependencyUUIDs, which has the added
advantage of adhering to spaces visibility. As a
consequence, rematerialization now correctly only
reloads the artifacts in the spaces that are visible
to the project that we are about to build.