Moving build.options.xxx to build.xxx
This is just the first step of a substantial refactoring
of buildOptions; the aim is using build.xxx only as a
set of defaults, that are then dissolved into the
regular project definitions.
Once that is done, we will no longer need to shuttle
around buildOptions everywhere, and the prototype of
all the relevant methods will be changed accordingly.