Adding a full options parsing infrastructure
We can now use "dbuild --debug" or "dbuild -d", getting
a complete help with "dbuild --help", and we have the
infrastructure needed to add further options if needed.

A small catch is that the sbt launcher steals options
from the command line, and there is no way to prevent
that. For instance "--version" is always intercepted
by the sbt launcher, which prints its own version rather
than passing "--version" to dbuild so that IT can print
its version. :-/

A possible approach could be to create a custom launcher,
identical to the standard one but just with a different
"main()" which does not unnecessarily intercept parts
of the command line.