Use a separate .m2 local repo for each Scala build
The ant build script calls maven directly, meaning that if we
use strange version numbers for the artifacts they
1) pollute ~/.m2, and
2) once one is in the cache, all hope to get it evicted is lost

Therefore: we need a local .m2 cache for each build, to avoid
pollution & collisions.

The cleanest approach would be to have a direct support for
.m2/repository customization within build.xml, for example
by modifying "localRepository" inside the 'artifact:dependencies'
sections when -Dmaven.repo.local is specified.

Lacking that support, the only other option is relying on the
(undocumented) ``_JAVA_OPTIONS`` environment variable, which is
the only way to actually force ant (and maven) to use a different
location for their .m2 cache.

Also, some minor formatting.