Add "init()" call to Git support
Currently a new repository will be initialized with a clone,
followed by a fetch in order to update it and set all the refs
correctly.

In principle the pair clone/fetch could be replaced by a
git init, then git remote add origin, then fetch. In the case
of remote repositories, that avoids one fetch of remote refs
from the remote repo.

In practice, the cloning is only done once from the remote
repositories, thanks to the built-in dbuild cache. Also,
the pair init+fetch is less efficient than clone+fetch in
the case of local repositories: clone will use hard links
(making things faster and using less disk space), whereas
init+fetch will copy the packs on the local filesystem.

Still, the init() call could be of use at some point, so
it is added here to the support/git/Git.scala file.