Replace "git checkout" with "git reset" --hard
This commit fixes an obscure and a bit tricky issue which
may cause incorrect updates of the git checkouts during
extraction.

In detail, the "git fetch" with the complex refspec used
in dbuild will stealthily update HEAD without updating
the checkout out files. A subsequent "git checkout"
interprets this condition as "files were changed",
and will therefore refuse to update the corresponding
files. Ouch.

That may affect extraction, which would see stale data,
but not the actual building stage, since the updated
HEAD enters into the UUID of the build directory, and
each successful build takes place in a new directory
in any case.

This commit addresses this bizarre issue by replacing
the "git checkout" with a more imperative
"git reset --hard", which always updates the files,
no matter what. Notably, the JGit implementation
was already doing that, because of different reasons.