Long overdue consolidation of autowrappings in the config file
dbuild uses a special "auto-wrapping" feature for many of its
structures, by which a single string, for example, may appear
where an array of strings is expected, or a single notification
may appear in place of a string of notifications, and so on.

dbuild uses a number of custom serializer/deserializers in
order to implement a suitable autowrapping.

This commit consolidates many of those routines into a single
generic version, which can more easily be recycled for other
structures as well.