The rewiring stage should be ok (the onLoad juggling is a bit of a hack)
The artifacts should now be properly rematerialized and, if everything
works as intended, the rewiring should take place automatically in
all levels of an sbt build.

There is some hackish behavior around onLoad, however. The main issue
is that we need to change the state in the upper levels of the build.
The state can only be changed by a command, or by onLoad: a simple
task won't do, and a command cannot be called from a task (that is,
we probably can extract the state and run a transformation, but
the resulting state is then lost, from within a task).

In order to transform the state in the upper levels, onLoad calls
directly a state-to-state transformation. The transformation builds
a new state, but then must reload it... which causes a new invocation
of onLoad, causing an endless sequence.

As a workaround, the onLoad handler restores its previous value before
reloading the new state, which seems to be working ok in the end.

This approach might potentially have drawbacks. There might be global
state involved, and the recursive onLoad call (there are two nested
calls to it, during execution) might be fragile and cause problems.
If we can come up with a better solution, so much the better.

Note: the code invoked by onLoad is functionally equivalent to an
sbt command.