"dbuild checkout" is complete, and seems to work properly
It is now possible to reproduce the compilation process that
dbuild performs, in all details, completely within a comfortable
new sbt session. This works much better than the old sbt plugin,
as it is now possible to really reproduce the dbuild compilation
mechanism.

Also, it is no longer necessary to fiddle with local/global
plugins settings, as "dbuild checkout" automatically prepares
everything, in a totally self-contained testing directory.

Documentation is still missing, but the options are listed
by issuing "dbuild --help".