Provision for dependency calculation for dependsOn(uri) projects
In the case of a dependent project defined implicitly using the
form dependsOn(uri), the project does not appear within
extracted.currentUnit.defined; that caused an exception while
trying to process the dependency graph. We now scan the list
of projects defined in extracted.structure.allProjects, and
add an empty dependency list if the project is not found in
extracted.currentUnit.defined.